requirements.txt
chromadb==0.3.25
langchain==0.0.218
openai==0.27.7
pdf2image==1.16.3
python-dotenv==1.0.0
python-magic==0.4.27
python-magic-bin==0.4.14
requests==2.28.0
tabulate==0.9.0
tiktoken==0.4.0
unstructured==0.7.1
